+++ react/tic-tac-toe/src/app/screens/Game/index.js
@@ -5,26 +5,17 @@ import { connect } from 'react-redux';
 import styles from './styles.module.scss';
 import Board from './components/Board';
 import MatchHistory from './components/MatchHistory';
-import { calculateWinner } from "~utils";
-import gameActionsCreator from "~redux/game/actions";
-import { JUMP_TEXT, GAME_START_TEXT, NEXT_PLAYER, WINNER, TIC, TAC } from "~constants";
+import JumpText from './components/JumpText';
+import { calculateWinner } from '~utils';
+import gameActionsCreator from '~redux/game/actions';
+import { NEXT_PLAYER, WINNER, TIC, TAC } from '~constants';
 
 class Game extends Component {
-  jumpText = move => {
-    return move ? JUMP_TEXT + move : GAME_START_TEXT;
-  }
-
   render() {
-    const { current, jumpTo, xIsNext } = this.props;
+    const { current, xIsNext } = this.props;
     const history = this.props.history;
     const winner = calculateWinner(current.squares);
-    const moves = history.map((step, move) => {
-      return (
-        <li key={move}>
-          <button onClick={() => jumpTo(move)}>{this.jumpText(move)}</button>
-        </li>
-      );
-    });
+    const moves = history.map((step, move) => <JumpText move={move} />);
     let status = null;
     if (winner) {
       status = WINNER + winner;
@@ -45,14 +36,14 @@ class Game extends Component {
   }
 }
 
-const mapStateToProps = ({ game: { history,stepNumber, xIsNext, current } }) => ({
-  history: history,
-  stepNumber: stepNumber,
-  xIsNext: xIsNext,
-  current: current
-  });
-const mapDispatchToProps = dispatch => ({
-  jumpTo: historyId => dispatch(gameActionsCreator.historyClick(historyId))
+const mapStateToProps = ({ game: { history, stepNumber, xIsNext, current } }) => ({
+  history,
+  stepNumber,
+  xIsNext,
+  current
 });
 
-export default connect(mapStateToProps, mapDispatchToProps)(Game);
+export default connect(
+  mapStateToProps,
+  null
+)(Game);
